Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/eventsource/dedicated-worker/eventsource-constructor-non-same-origin-expected.txt

Blocked access to external URL https://example.not/test
Blocked access to external URL http://example.not/

FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (http://example.not/) assert_equals: source.readyState expected 2 but got 0
FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (https://example.not/test) assert_equals: source.readyState expected 2 but got 0
FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (ftp://example.not/) assert_equals: source.readyState expected 2 but got 0
FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (about:blank) assert_equals: source.readyState expected 2 but got 0
FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (mailto:whatwg@awesome.example) assert_equals: source.readyState expected 2 but got 0
FAIL dedicated worker - EventSource: constructor (act as if there is a network error) (javascript:alert('FAIL')) assert_equals: source.readyState expected 2 but got 0
